add Pipe connection

+++ Basalaev.Daniil/laba2/Library/src/Conn/conn_fifo.cpp
@@ -9,6 +9,7 @@
 std::unique_ptr<ConnFifo> ConnFifo::crateHostFifo(std::string const& fifoPath)
 {
     ConnFifo* fifo = new ConnFifo();
+    fifo->m_isHost = true;
     fifo->m_fifoPath = fifoPath;
 
     if (mkfifo(fifo->m_fifoPath.c_str(), 0666) == -1)
@@ -41,6 +42,7 @@ std::unique_ptr<ConnFifo> ConnFifo::crateHostFifo(std::string const& fifoPath)
 std::unique_ptr<ConnFifo> ConnFifo::crateClientFifo(std::string const& fifoPath)
 {
     ConnFifo* fifo = new ConnFifo();
+    fifo->m_isHost = false;
     fifo->m_fifoPath = fifoPath;
 
     fifo->m_readFileDisriptor = open(fifo->m_fifoPath.c_str(), O_RDONLY | O_NONBLOCK);
@@ -78,18 +80,29 @@ ConnFifo::~ConnFifo()
         close(m_writeFileDisriptor);
     }
 
-    if (!m_fifoPath.empty())
+    if (!m_fifoPath.empty() && m_isHost)
     {
         unlink(m_fifoPath.c_str());
     }
 }
 
-bool ConnFifo::Read(void* buf, size_t count)
+bool ConnFifo::Read(void* buf, size_t maxSize)
 {
     if (!isValid()) { return false; }
 
-    ssize_t bytesRead = read(m_readFileDisriptor, buf, count);
+    char tmpBuf[alias::MAX_MSG_SIZE];
+    ssize_t bytesRead = read(m_readFileDisriptor, tmpBuf, maxSize);
     if (bytesRead == -1) { return false; }
+
+    if ((tmpBuf[0] == '0' && m_isHost) || (tmpBuf[0] == '1' && !m_isHost))
+    {
+        std::memcpy(buf, &tmpBuf[1], --bytesRead);
+    }
+    else
+    {
+        Write(&tmpBuf[1], bytesRead - 1);
+        return false;
+    }
 
     return bytesRead > 0;
 }
@@ -98,8 +111,11 @@ bool ConnFifo::Write(const void* buf, size_t count)
 {
     if (!isValid()) { return false; }
 
-    ssize_t bytesWritten = write(m_writeFileDisriptor, buf, count);
+    char tmpBuf[alias::MAX_MSG_SIZE];
+    tmpBuf[0] = m_isHost ? '1' : '0';
+    std::memcpy(&tmpBuf[1], buf, count);
+    ssize_t bytesWritten = write(m_writeFileDisriptor, tmpBuf, count + 1);
     if (bytesWritten == -1) { return false; }
 
-    return static_cast<size_t>(bytesWritten) == count;
+    return static_cast<size_t>(bytesWritten) == count + 1;
 }
